requirements.txt
asgiref==3.2.10
Django==3.1.1
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-filter==2.4.0
django-imagekit==4.0.2
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==2.1.2
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
Markdown==3.3
markdown2==2.3.9
pilkit==2.0
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.2.0
